back_end/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.27
msal==1.26.0
requests==2.31.0
gunicorn==21.2.0 